+++ cpp/brain/controller/ExtendedANNWeights.cpp
@@ -164,7 +164,7 @@ ExtNNController::writeNetwork(std::ofstream &write_to)
     }
   }
   std::string *names = new std::string[allNeurons_.size()];
-  for (int i = 0; i < allNeurons_.size(); i++) {
+  for (size_t i = 0; i < allNeurons_.size(); i++) {
     std::stringstream nodeName;
     nodeName << allNeurons_[i]->Id() + " of type: " + allNeurons_[i]->getType() << std::endl;
     for (std::pair<std::string, double> param : allNeurons_[i]->getNeuronParameters()) {
